+++ bot/commands/debug.js
@@ -7,6 +7,7 @@ module.exports = {
     description: 'toggle debug mode.',
     args: false,
     usage: '',
+    example: '',
     useApi: true,
     superuserOnly: true,
 	execute: async (message, args, config, client) => {