+++ bot/commands/ping.js
@@ -2,7 +2,8 @@ module.exports = {
 	name: 'ping',
     description: 'Ping!',
     args: false,
-    usage: '',
+	usage: '',
+	example: '',
 	execute(message, args, config, client) {
 		message.channel.send('Pong.');
 	},